Changes made in change form

+++ src/app/pages/changes/changes.presenter.ts
@@ -2,6 +2,7 @@ import { Component, OnInit } from '@angular/core';
 import { ChangesComponent } from "./changes.component";
 import { ChangeModel } from '../../services/model-interface/interfaces';
 import { ModelService } from '../../services/model-services/model.service';
+import Swal from 'sweetalert2';
 
 @Component({
   selector: 'app-changes-presenter',
@@ -30,12 +31,37 @@ export class ChangesPresenter implements OnInit {
   }
 
   deleteChangeById(id: number){
-     this.modelService.deleteChange(id).subscribe({
-      next: (data) => {
-        console.log("Deleted successfully", data)
-      },
-      error: (err) => console.log('Delete Error :', err)
-     })
+      Swal.fire({
+               title: 'Are you sure?',
+               text: 'You want to delete this Change',
+               icon: 'warning',
+               showCancelButton: true,
+               confirmButtonColor: '#d33',
+               cancelButtonColor: '#3085d6',
+               confirmButtonText: 'Delete'
+             }).then((result) => {
+               if (result.isConfirmed) {
+                 this.modelService.deleteChange(id).subscribe({
+                   next: (data) => {
+                     console.log("Deleted Change log:", data);
+                     Swal.fire(
+                       'Deleted!',
+                       'The Change has been deleted.',
+                       'success'
+                     ).then(() => {
+                       location.reload(); // Reload the page after the alert is closed
+                     });
+                   },
+                   error: (err) => {
+                     Swal.fire(
+                       'Error!',
+                       'Something went wrong while deleting.',
+                       'error'
+                     );
+                   }
+                 });
+               }
+             });
   }
 
 }
